herccia: Utilities for creating Civic Infrastructure Assessment reports

Perform the data tasks for a StriveTogether Civic Infrastructure Assessment. Initially, this package focuses on public data from the US Census and Wisconsin's Department of Public Instruction, and internal data from Racine Unified School District.

See also

Author

Maintainer: Ben Taft ben.taft@career2cradle.org

Other contributors:

Higher Expectations for Racine County [copyright holder]